projects/file-server/requirements.txt
fastapi>=0.95.0,<1.0.0
uvicorn[standard]>=0.22.0,<1.0.0
aiofiles>=23.1.0,<24.0.0
python-multipart>=0.0.6,<0.1.0
pillow>=10.0.0,<11.0.0
python-magic>=0.4.27,<0.5.0
aiohttp>=3.8.4,<4.0.0
pydantic>=1.10.8,<2.0.0
python-dotenv>=1.0.0,<2.0.0
validators>=0.20.0,<0.21.0
aiokafka>=0.8.1,<0.9.0
python-jose>=3.3.0,<4.0.0
requests>=2.31.0,<3.0.0
